move components to one folder
